# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Live Test API
  version: 0.0.1
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/users:
    post:
      tags:
        - User
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: Duplicate email or nickname
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
...
components:
  schemas:
    ApiEnvelope:
      type: object
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
          description: Successful response payload.
        error:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - fieldErrors
        - timestamp
      properties:
        code:
          type: string
          example: COMMON-001
        message:
          type: string
          example: 잘못된 입력입니다.
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorResponse'
        timestamp:
          type: string
          format: date-time
          example: '2026-07-27T05:02:11Z'
    FieldErrorResponse:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: message
        message:
          type: string
          example: must not be blank
    CreateUserRequest:
      type: object
      required:
        - email
        - nickname
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: user@example.com
        nickname:
          type: string
          minLength: 1
          maxLength: 50
          example: live-user
...